Close fd when acquire fails.
Currently, the lockfile's fd is opened when attempting to acquire the
lock, and remains open until the lock is released. This means that in
cases where the lock is not successfully acquired before the objects go
away, a ResourceWarning of an unclosed fd is printed. This can also
cause unclosed fds to accumulate, or even cause a process to run out of
fds.

This came up for me when attempting to clean up an old tmp/lockdir, when
attempting to acquire many locks but immediately give up when
unsuccessful, and then remove them when successful.

This closes the lockfile fd when acquire fails to actually acquire the
lock.