reintoduce current_thread_functor
This change reintoruced the current_thread_functor to the
ReaderWriterLock __init__. This parmater now default to None
and if not passed will result in threading.current_thread being
used to detect if its safe for the current execution context to
renter a lock. If you are using fasterns ReaderWriterLocks in a
context where eventlet or other user land threads are in use, a
alternitive functor should be provided.

Closes: #96